Move version check/update utilities into a dedicated package.
